requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=2.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-generativeai>=0.3.0 